requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
helena-core==0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.3.5
SQLAlchemy==2.0.40
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
wssai==0.1
pytesseract
opencv-python
Pillow
numpy
scikit-image
scikit-learn
scipy
tqdm
tqdm==4.65.0
tqdm
tqdm==4.65.0
tqdm
tqdm==4.65.0